scripts/requirements.txt
mysql-connector-python==8.2.0
pymongo==4.6.0
faker==22.0.0
